requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Scientific computing
scipy>=1.9.0
scikit-learn>=1.1.0

# Optimization
scipy>=1.9.0

# Statistical analysis
statsmodels>=0.13.0

# Plotting and visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# R integration for Q-density estimation
rpy2>=3.5.0

# Parallel processing
joblib>=1.1.0

# Configuration management
PyYAML>=6.0

# Interpolation and mathematical functions
scipy>=1.9.0

# Progress bars
tqdm>=4.64.0

# File path handling
pathlib2>=2.3.0

# Date handling
python-dateutil>=2.8.0

# Excel file support
openpyxl>=3.0.0
xlrd>=2.0.0

# Optional: Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Advanced clustering and dimensionality reduction
umap-learn>=0.5.0

# Optional: Advanced plotting
plotly>=5.0.0 